+++ src/filters/GeometricMapFilter.js
@@ -45,6 +45,8 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
     ,matrix: null
     ,centerX: 0
     ,centerY: 0
+    ,dx: 0
+    ,dy: 0
     ,angle: 0
     ,radius: 0
     ,wavelength: 0
@@ -68,6 +70,8 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
         self.matrix = null;
         self.centerX = null;
         self.centerY = null;
+        self.dx = null;
+        self.dy = null;
         self.angle = null;
         self.radius = null;
         self.wavelength = null;
@@ -94,6 +98,8 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
                 ,matrix: self.matrix
                 ,centerX: self.centerX
                 ,centerY: self.centerY
+                ,dx: self.dx
+                ,dy: self.dy
                 ,angle: self.angle
                 ,radius: self.radius
                 ,wavelength: self.wavelength
@@ -121,6 +127,8 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
             self.matrix = params.matrix;
             self.centerX = params.centerX;
             self.centerY = params.centerY;
+            self.dx = params.dx;
+            self.dy = params.dy;
             self.angle = params.angle;
             self.radius = params.radius;
             self.wavelength = params.wavelength;
@@ -249,6 +257,15 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
         return self;
     }
     
+    ,shift: function( dx, dy ) {
+        var self = this;
+        self.dx = (dx!==undef) ? dx : 0; 
+        self.dy = (dy!==undef) ? dy : self.dx; 
+        self._mapName = "shift";  
+        self._map = Maps.shift; 
+        return self;
+    }
+    
     ,reset: function( ) {
         var self = this;
         self._mapName = null; 
@@ -278,7 +295,8 @@ var GeometricMapFilter = FILTER.GeometricMapFilter = FILTER.Class( FILTER.Filter
         return this._isOn && this._map;
     }
 });
-
+// aliases
+GeometricMapFilter.prototype.translate = GeometricMapFilter.prototype.shift;
 
 //
 //
@@ -367,6 +385,25 @@ Maps = {
         return dst;
     }
 
+    ,"shift": function( self, im, w, h ) {
+        var x, y, yw, i, j, l=im.length, dst=new IMG(l),
+            dx = -self.dx, dy = -self.dy;
+        
+        if ( dx < 0 ) dx += w;
+        if ( dy < 0 ) dy += h;
+        
+        x=0; y=0; yw=0;
+        for (i=0; i<l; i+=4, x++)
+        {
+            if (x>=w) { x=0; y++; yw+=w; }
+            
+            j = ((x+dx)%w + ((y+dy)%h)*w)<<2;
+            dst[i] = im[j];   dst[i+1] = im[j+1];
+            dst[i+2] = im[j+2];  dst[i+3] = im[j+3];
+        }
+        return dst;
+    }
+    
     ,"flipX": function( self, im, w, h ) {
         var x, y, yw, i, j, l=im.length, dst=new IMG(l);